+++ services/cert_manager_standard/cert_manager_engine/main/core/src/cert_manager_uri.c
@@ -549,7 +549,7 @@ int32_t CertManagerGetUidFromUri(const struct CmBlob *uri, uint32_t *uid)
         return CMR_ERROR_INVALID_ARGUMENT;
     }
 
-    *uid = atoi(uriObj.app);
+    *uid = (uint32_t)atoi(uriObj.app);
     (void)CertManagerFreeUri(&uriObj);
     return CM_SUCCESS;
 }
